+++ CHANGELOG.md
@@ -21,11 +21,13 @@
 ### CHANGED
 - Preserve `Eloquent\Builder` when overriding the default ordering of dataTables when using `EloquentEngine`.
 - Preserve `Eloquent\Builder` when using filterColumn api. Allows us to use model scope and any eloquent magics.
+- Fractal integration extracted to own plugin [laravel-datatables-fractal](https://github.com/yajra/laravel-datatables-fractal).
 
 ### REMOVED
 - Remove filterColumn api magic query method in favor of closure.
 - Remove support on older snake_case methods.
 - Remove silly implementation of proxying query builder calls via magic method. 
+- Removed unused methods.
 
 ### FIXED
 - Fix #1068.